requirements.txt
python-telegram-bot
google-generativeai
fastapi
uvicorn
python-multipart